변수 추가 및 코드 수정
조건문이 복잡해서 변수로 분리해주고 코드를 수정했다.
